requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary>=2.9.10
Flask-Migrate==4.0.5
gunicorn==21.2.0
